+++ docs/changelog.md
@@ -14,6 +14,7 @@ Click on the version in the following list to see all changes since the previous
 
 * chore: drop support for Jira 8.13.x and ensure support for latest versions
 * feat: allow users to edit existing user templates (#51)
+* feat: validate syntax when creating and editing user templates (#42)
 
 ### [22.11.0] - 2022-11-14